+++ js/screens/game/game-view.js
@@ -3,8 +3,6 @@ import ArtistView from './artist-view.js';
 import GenreView from './genre-view.js';
 import HeaderView from './header-view.js';
 
-const ARTIST_LVL_COUNT = 9;
-
 const update = (container, view) => {
   while (container.firstChild) {
     container.firstChild.remove();
@@ -30,7 +28,7 @@ export default class GameView extends AbstractView {
 
   updateScreen() {
     this.updateHeader();
-    const view = (this.model.data.gameData.stat.length < ARTIST_LVL_COUNT) ? (new ArtistView(this.model.gameQuestion)) : (new GenreView(this.model.gameQuestion));
+    const view = (this.model.data.gameData.stat.length < this.model.data.state.artistScreens) ? (new ArtistView(this.model.gameQuestion)) : (new GenreView(this.model.gameQuestion));
     view.playerHandler = (artistPlayer, playerControl) => this.model.playerHandler(artistPlayer, playerControl);
     view.artistAnswerHandler = (evt, artistAnswers) => this.model.artistAnswerHandler(evt, artistAnswers);
     view.playersHandler = (evt, genrePlayers, playersControls) => this.model.playersHandler(evt, genrePlayers, playersControls);
