--- js/utils/show-element.js
@@ -1,11 +1,0 @@
-const app = document.querySelector(`.app`);
-const mainSection = app.querySelector(`.main`);
-
-const showScreen = (screen) => {
-  while (mainSection.firstChild) {
-    mainSection.firstChild.remove();
-  }
-  mainSection.appendChild(screen.element);
-};
-
-export {showScreen};
